support hub lora
