+++ docs/source/Instruction/命令行参数.md
@@ -268,7 +268,7 @@ Vera使用`target_modules`, `target_regex`, `modules_to_save`三个参数.
 - enforce_eager: vllm使用pytorch eager模式还是建立cuda graph. 默认为`False`. 设置为True可以节约显存, 但会影响效率.
 - 🔥limit_mm_per_prompt: 控制vllm使用多图, 默认为`None`. 例如传入`--limit_mm_per_prompt '{"image": 10, "video": 5}'`
 - vllm_max_lora_rank: 默认为`16`. vllm对于lora支持的参数
-- lora_modules: 用于支持多lora动态切换，默认为`[]`
+- lora_modules: 用于支持部署时多lora动态切换，默认为`[]`
 
 
 ### 合并参数
