+++ swift/llm/argument/deploy_args.py
@@ -41,6 +41,7 @@ class DeployArguments(InferArguments):
 
     def __post_init__(self):
         super().__post_init__()
+        self._parse_lora_modules()
         self.port = find_free_port(self.port)
 
     def _init_stream(self):
@@ -54,3 +55,16 @@ class DeployArguments(InferArguments):
             return
         self.result_path = self.get_result_path('deploy_result')
         logger.info(f'args.result_path: {self.result_path}')
+
+    def _parse_lora_modules(self) -> None:
+        from swift.llm import LoRARequest
+        if len(self.lora_modules) == 0:
+            self.lora_request_list = []
+            return
+        assert self.infer_backend in {'vllm', 'pt'}
+        lora_request_list = []
+        for i, lora_module in enumerate(self.lora_modules):
+            lora_name, lora_path = lora_module.split('=')
+            lora_path = self._get_ckpt_dir(lora_path)
+            lora_request_list.append(LoRARequest(lora_name, lora_path))
+        self.lora_request_list = lora_request_list
