+++ swift/llm/argument/infer_args.py
@@ -6,7 +6,7 @@ from typing import List, Literal, Optional
 
 import torch.distributed as dist
 
-from swift.llm import LoRARequest, get_template_meta
+from swift.llm import get_template_meta
 from swift.utils import get_logger, is_dist
 from .base_args import BaseArguments, to_abspath
 from .base_args.model_args import ModelArguments
@@ -73,7 +73,7 @@ class VllmArguments:
     limit_mm_per_prompt: Optional[str] = None  # '{"image": 10, "video": 5}'
     vllm_max_lora_rank: int = 16
 
-    lora_modules: List[str] = field(default_factory=list)
+    lora_modules: List[str] = field(default_factory=list)  # for deploy
 
     def __post_init__(self):
         self.limit_mm_per_prompt = ModelArguments.parse_to_dict(self.limit_mm_per_prompt)
@@ -151,12 +151,13 @@ class InferArguments(MergeArguments, VllmArguments, LmdeployArguments, BaseArgum
 
     def __post_init__(self) -> None:
         if self.ckpt_dir:
-            self._load_ckpt_dir()
+            self.ckpt_dir = self._get_ckpt_dir(self.ckpt_dir)
+            self.load_args_from_ckpt(self.ckpt_dir)
+
         self._init_weight_type(self.ckpt_dir)
         BaseArguments.__post_init__(self)
         MergeArguments.__post_init__(self)
         VllmArguments.__post_init__(self)
-        self._parse_lora_modules()
 
         self._init_result_path()
         self._init_eval_human()
@@ -170,14 +171,3 @@ class InferArguments(MergeArguments, VllmArguments, LmdeployArguments, BaseArgum
             eval_human = False
         self.eval_human = eval_human
         logger.info(f'Setting args.eval_human: {self.eval_human}')
-
-    def _parse_lora_modules(self) -> None:
-        if len(self.lora_modules) == 0:
-            self.lora_request_list = []
-            return
-        assert self.infer_backend in {'vllm', 'pt'}
-        lora_request_list = []
-        for i, lora_module in enumerate(self.lora_modules):
-            lora_name, lora_path = lora_module.split('=')
-            lora_request_list.append(LoRARequest(lora_name, lora_path))
-        self.lora_request_list = lora_request_list
