+++ swift/llm/infer/deploy.py
@@ -63,12 +63,16 @@ class SwiftDeploy(SwiftInfer):
         finally:
             self._compute_infer_stats()
 
-    async def get_available_models(self):
+    def _get_model_list(self):
         args = self.args
         model_list = [args.served_model_name or args.model_suffix]
         if args.lora_request_list is not None:
             model_list += [lora_request.lora_name for lora_request in args.lora_request_list]
-        data = [Model(id=model_id, owned_by=args.owned_by) for model_id in model_list]
+        return model_list
+
+    async def get_available_models(self):
+        model_list = self._get_model_list()
+        data = [Model(id=model_id, owned_by=self.args.owned_by) for model_id in model_list]
         return ModelList(data=data)
 
     async def _check_model(self, request: ChatCompletionRequest) -> Optional[str]:
@@ -156,6 +160,7 @@ class SwiftDeploy(SwiftInfer):
 
     def run(self):
         args = self.args
+        logger.info(f'model_list: {self._get_model_list()}')
         uvicorn.run(
             self.app, host=args.host, port=args.port, ssl_keyfile=args.ssl_keyfile, ssl_certfile=args.ssl_certfile)
